Create README.md
This is a SQL based project.
In this project, I analyzed the Instagram database by importing the database to MYSQL, and solving a set of questions using joins, group by, having clause, etc. The key highlights of this project were to find the activity of various instagram users such as number of active users, average number of photos posted, most liked photos, most used hashtags, number of inactive users, etc.